doze: allow grayscale even if invert boolean is false
for dark themes, we are setting the config boolean for inverting
doze notifications to false.  in addition to preventing
color inversion, this was also preventing the notification
from grayscaling.  as a result, on dark themes (boolean false), we
were seeing color icons on doze notifications.  this commit fixes
the grayscaling, and brings back the very aesthetically pleasing
(imo) grayscale-to-color fade-in for the notifications' app icons.

Change-Id: Ifc5efbccbeb02910684b76793721b10f1d64f870

# Conflicts:
#	packages/SystemUI/src/com/android/systemui/ViewInvertHelper.java